+++ frontend/src/router/index.js
@@ -2,7 +2,7 @@ import { createRouter, createWebHistory } from 'vue-router'
 import HomeView from '../views/HomeView.vue'
 import ConsultSubjectsView from '@/views/ConsultSubjectsView.vue'
 import DashboardView from '@/views/DashboardView.vue'
-import FriendsScreen from '@/views/FriendsScreen.vue'
+import FriendsScreen from '@/views/FriendsView.vue'
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
