+++ frontend/src/views/FriendsView.vue
